+++ CHANGELOG.md
@@ -17,7 +17,7 @@ and this project will adhere to [Semantic Versioning](https://semver.org/spec/v2
 
 ### Changed
 
-- ⚠️ BREAKING CHANGE: In NPM package - typescript types are now exported by default
+- ⚠️ BREAKING CHANGE: NPM package - typescript types are now exported by default
 - add `querystring` and `params` to schemas
 
 ### Added